move emf handling to its own utility class
